requirements.txt
streamlit
chromadb
zhipuai
python-dotenv
pymupdf4llm